requirements.txt
# trading_ai_monitor_v2/requirements.txt
python-telegram-bot==20.7
telethon==1.28.5
python-dotenv==1.0.0
pandas==2.0.3
numpy>=1.21.0
scipy>=1.7.0
ta==0.10.2
requests==2.31.0
aiohttp==3.8.5